Added a whole lot of Stubs to implement a simple(ish) Storage implementation that works without a postgresql database.
Implemented a test that uploads and publishes a module in one go. Or rather,
fixed the one that was there. This could be better, I basically gave up chasing
the permissions issue, because it works perfectly outside of the test
environment, all roles and things suggest that it should work, and it simply
took too long to figure it out. So for the moment we grab Manager rights to
do publishing.